Timus 1022

+++ 18 HW Timus 1022/Timus_1022.cpp
@@ -64,7 +64,7 @@ private:
     int size;
     vector<vector<int>> gr;
     bool* used;
-    stack<short> res;
+    stack<int> res;
 };
 
 int main()
